api/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.7
importlib_resources==6.4.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
SQLAlchemy==2.0.32
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
